redo addRecord, removeRecord, processRecord, and traceRecord
