+++ src/Makefile
@@ -23,11 +23,6 @@ INC += pv/pvDatabase.h
 
 INC += pv/channelProviderLocal.h
 
-#INC += pv/traceRecord.h
-#INC += pv/removeRecord.h
-#INC += pv/addRecord.h
-#INC += pv/processRecord.h
-
 INC += pv/pvSupport.h
 INC += pv/controlSupport.h
 INC += pv/scalarAlarmSupport.h
