+++ src/special/Makefile
@@ -2,20 +2,15 @@
 
 SRC_DIRS += $(PVDATABASE_SRC)/special
 
-#LIBSRCS += traceRecord.cpp
-#LIBSRCS += removeRecord.cpp
-#LIBSRCS += addRecord.cpp
-#LIBSRCS += processRecord.cpp
+DBD += traceRecordRegister.dbd
+DBD += removeRecordRegister.dbd
+DBD += addRecordRegister.dbd
+DBD += processRecordRegister.dbd
 
-#DBD += traceRecordRegister.dbd
-#DBD += removeRecordRegister.dbd
-#DBD += addRecordRegister.dbd
-#DBD += processRecordRegister.dbd
-
-#LIBSRCS += traceRecordRegister.cpp
-#LIBSRCS += removeRecordRegister.cpp
-#LIBSRCS += addRecordRegister.cpp
-#LIBSRCS += processRecordRegister.cpp
+LIBSRCS += traceRecordRegister.cpp
+LIBSRCS += removeRecordRegister.cpp
+LIBSRCS += addRecordRegister.cpp
+LIBSRCS += processRecordRegister.cpp
 
 DBD += pvdbcrTraceRecordRegister.dbd
 DBD += pvdbcrRemoveRecordRegister.dbd
